% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{address}
\alias{address}
\title{address}
\usage{
address(data.path, fresh = FALSE)
}
\arguments{
\item{data.path}{character vector}

\item{fresh}{logical TRUE for fresh run}
}
\description{
Creates and loads points.address file
}
\keyword{clean,}
\keyword{geocode}
\keyword{points,}
