% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\name{shapes.boulder}
\alias{shapes.boulder}
\title{shapes.boulder}
\usage{
shapes.boulder(l.path, fresh = FALSE)
}
\arguments{
\item{l.path}{character vector}

\item{fresh}{logical TRUE for fresh run}
}
\description{
Not currently running!
}
\keyword{clean,}
\keyword{outlines,}
\keyword{shapes}
