% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\name{shapes.denver}
\alias{shapes.denver}
\title{shapes.denver}
\usage{
shapes.denver(l.path, fresh = FALSE)
}
\arguments{
\item{l.path}{character vector package data path}

\item{fresh}{logical TRUE for fresh run}
}
\description{
loads denver oulines as spatialPolygonDataFrame
}
\keyword{clean,}
\keyword{outlines,}
\keyword{shapes}
