% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\name{shapes}
\alias{shapes}
\title{shapes}
\usage{
shapes(data.path, fresh = FALSE)
}
\arguments{
\item{data.path}{character vector}

\item{fresh}{logical TRUE for fresh run}
}
\description{
currently only returns denver building footprint shapes
}
\keyword{clean,}
\keyword{outlines,}
\keyword{shapes}
