% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\name{address}
\alias{address}
\title{address}
\usage{
address(data.path, fresh = FALSE)
}
\arguments{
\item{data.path}{character vector}

\item{fresh}{logical TRUE for fresh run}
}
\description{
Creates and loads outlines.address file
}
\keyword{clean}
\keyword{outliens,}
