% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{address.update}
\alias{address.update}
\title{address.update}
\usage{
address.update(dt.pkg.data, points.address.new,
  points.source = "geocode.points", points.address)
}
\arguments{
\item{dt.pkg.data}{full package data data.table}

\item{points.address.new}{new points}

\item{points.source}{geocoded points}

\item{points.address}{the original points.address files}
}
\description{
Creates/loads and updates points.address file
}
\keyword{address,}
\keyword{clean,}
\keyword{points,}
\keyword{update}
