% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlines.R
\name{address.update}
\alias{address.update}
\title{address.update}
\usage{
address.update(data.path, outlines.address, outlines.address.new,
  outline.source = "points")
}
\arguments{
\item{data.path}{source path for the package}

\item{outlines.address}{the outlines.address data.table}

\item{outlines.address.new}{the new (geocoded outlines.address)}

\item{outline.source}{the source type for the outline (default to points here)}
}
\description{
updates outlines.address with point data
}
\keyword{outlines,}
\keyword{points,}
\keyword{update}
