% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{address.update}
\alias{address.update}
\title{address.update}
\usage{
address.update(data.path, points.address.new,
  points.source = "geocode.points", points.address)
}
\arguments{
\item{data.path}{character vector}

\item{points.address.new}{new points}

\item{points.source}{geocoded points}

\item{points.address}{the original points.address files}
}
\description{
Creates and loads outlines.address file
}
\keyword{clean}
\keyword{outliens,}
